+++ src/redash_iodide/settings.py
@@ -6,7 +6,6 @@ import os
 
 from redash.handlers.base import BaseResource
 from redash.permissions import require_permission
-from redash.settings.helpers import array_from_string
 
 from .resources import add_resource
 
@@ -21,26 +20,6 @@ IODIDE_NOTEBOOK_API_URL = os.environ.get("REDASH_IODIDE_NOTEBOOK_API_URL", "")
 
 # The auth token that this extension uses to create new Iodide notebooks
 IODIDE_AUTH_TOKEN = os.environ.get("REDASH_IODIDE_AUTH_TOKEN", "")
-
-# The JWT request header for authenticating Iodide requests
-IODIDE_JWT_AUTH_HEADER = os.environ.get("REDASH_IODIDE_JWT_AUTH_HEADER", "")
-
-# The shared secret key
-IODIDE_JWT_AUTH_KEY = os.environ.get("REDASH_IODIDE_JWT_AUTH_KEY", "")
-
-IODIDE_JWT_AUTH_ISSUER = os.environ.get(
-    "REDASH_IODIDE_JWT_AUTH_ISSUER", "https://iodide.io/"
-)
-
-IODIDE_JWT_AUTH_AUDIENCE = os.environ.get("REDASH_IODIDE_JWT_AUTH_AUDIENCE", "")
-
-IODIDE_JWT_AUTH_SUBJECT = os.environ.get(
-    "REDASH_IODIDE_JWT_AUTH_SUBJECT", "query:create"
-)
-
-IODIDE_JWT_AUTH_ALGORITHMS = array_from_string(
-    os.environ.get("REDASH_IODIDE_JWT_AUTH_ALGORITHMS", "ES256")
-)
 
 
 class IodideSettingsResource(BaseResource):